feat(cli): add a `token` subcommand to moq-cli
The generate/sign/verify command tree moves into `moq_token::cli::Args`
behind the crate's new `cli` feature. `moq-token-cli` flattens it, and
`moq-cli` nests it under `moq token`, so one implementation backs both
entry points instead of the CLI logic living in a binary nothing else
can reach.

`token` joins `devices` as a local verb: it runs before any transport is
bound and rejects `--client-connect` / `--server-bind` rather than
silently ignoring them.
