+++ rs/moq-token-cli/Cargo.toml
@@ -17,4 +17,4 @@ path = "src/main.rs"
 [dependencies]
 anyhow = "1"
 clap = { version = "4", features = ["derive"] }
-moq-token = { workspace = true }
+moq-token = { workspace = true, features = ["cli"] }
