+++ rs/moq-token/src/lib.rs
@@ -7,6 +7,8 @@
 
 mod algorithm;
 mod claims;
+#[cfg(feature = "cli")]
+pub mod cli;
 mod error;
 mod generate;
 mod key;
